+++ Development Project/Interview.Web/Controllers/InventoryController.cs
@@ -1,8 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Sparcpoint.Abstract;
+using Sparcpoint.Abstractions;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace Interview.Web.Controllers
