+++ Development Project/Interview.Web/Controllers/ProductController.cs
@@ -1,9 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Sparcpoint.Abstract;
-using Sparcpoint.Implementations;
-using System;
-using System.Collections.Generic;
-using System.Linq;
+using Sparcpoint.Abstractions;
 using System.Threading.Tasks;
 
 namespace Interview.Web.Controllers
