+++ Development Project/Sparcpoint.Core/Exceptions/ParameterRequiredException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Sparcpoint
+namespace Sparcpoint.Exceptions
 {
     public class ParameterRequiredException : Exception
     {
