+++ Development Project/Sparcpoint.Core/Models/CategoryCategories.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
+﻿using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Sparcpoint.Core.Models
 {
