+++ Development Project/Sparcpoint.SqlServer.Abstractions/Provider/OrderByClause.cs
@@ -5,7 +5,7 @@ namespace Sparcpoint.SqlServer.Abstractions
 {
     public class OrderByClause : IEnumerable<string>
     {
-        private readonly List<string> _Backer = new List<string>();
+        private readonly List<string> _Backer = new();
 
         public void Add(string columnName, OrderByDirection direction)
         {
